requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pyyaml==6.0.1
requests==2.31.0
swagger-spec-validator==3.0.3
jsonschema==4.19.2
Pillow==10.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
